create bottom view to order food

+++ app/src/main/java/com/sun_asterisk/myfood/data/local/dao/Converters.kt
@@ -26,7 +26,7 @@ class Converters {
 
     @TypeConverter
     fun fromJsonToLocation(value: String): ArrayList<Float>? {
-        val listType = object : TypeToken<ArrayList<Long>>() {}.type
+        val listType = object : TypeToken<ArrayList<Float>>() {}.type
         value.notNull {
             return Gson().fromJson(it, listType)
         }
